refactor(api): move the HttpClient interface into the published API
Reaching the network is the most common thing a plugin does, and it was the one
thing the API declined to describe. The interface lived in plugins/common, a
bundled helper module, so a third-party plugin either depended on something
that is not the published API or shipped its own HTTP stack inside its JAR.

Stating it in api/ is also the precondition for the application owning the
transport. A proxy setting is only reliable if plugins receive a client rather
than building their own, because a plugin cannot then opt out of it by
accident.

Nothing else changes. Same two methods, same signatures, same KtorHttpClient
behind them, so every call site keeps its code and the five plugin test fakes
keep compiling against the interface untouched.

One hazard the move introduced: KtorHttpClient star-imports io.ktor.client.*,
which contains Ktor's own HttpClient. Importing ours by name beats the star
import, so Ktor's client is now aliased rather than silently resolving to the
wrong type.
