+++ plugins/wikimedia-reference/src/main/kotlin/com/github/ahatem/qtranslate/plugins/wikimedia/WikimediaClient.kt
@@ -1,7 +1,7 @@
 package com.github.ahatem.qtranslate.plugins.wikimedia
 
 import com.github.ahatem.qtranslate.api.plugin.ServiceError
-import com.github.ahatem.qtranslate.plugins.common.HttpClient
+import com.github.ahatem.qtranslate.api.plugin.HttpClient
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
