+++ plugins/yandex-web-services/src/main/kotlin/com/github/ahatem/qtranslate/plugins/yandexweb/YandexWebClient.kt
@@ -1,7 +1,7 @@
 package com.github.ahatem.qtranslate.plugins.yandexweb
 
 import com.github.ahatem.qtranslate.api.plugin.ServiceError
-import com.github.ahatem.qtranslate.plugins.common.HttpClient
+import com.github.ahatem.qtranslate.api.plugin.HttpClient
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
