+++ plugins/yandex-web-services/src/test/kotlin/com/github/ahatem/qtranslate/plugins/yandexweb/YandexWebTranslatorServiceTest.kt
@@ -3,7 +3,7 @@ package com.github.ahatem.qtranslate.plugins.yandexweb
 import com.github.ahatem.qtranslate.api.language.LanguageCode
 import com.github.ahatem.qtranslate.api.plugin.ServiceError
 import com.github.ahatem.qtranslate.api.translator.TranslationRequest
-import com.github.ahatem.qtranslate.plugins.common.HttpClient
+import com.github.ahatem.qtranslate.api.plugin.HttpClient
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
 import com.github.michaelbull.result.Result
